requirements.txt
pyyaml
pandas
sklearn
scipy
scikit-image
matplotlib
seaborn
tqdm
albumentations
tensorboard